Lab 4B: copy-on-write fork
A user-level fork that calls syscall fork() then copies parent env's
virtual address space mapping, allocates child's user exception stack,
sets up its page fault handler and sets it as RUNNABLE.

A new duppage that doesn't copy the content of a page, but the mapping.

A user-level page fault handler that does copy-on-write when writing
into a copy-on-write page causes a page fault.