Lab 5: serve write
Implement file write.
On the client side of the PRC, populate IPC page with write arguments
and data to write, then call IPC.
On the file server side of the PRC, call file_write with the arguments
passed from the IPC and advance the FD offset.

Fix file read: advance the FD offset after each read.